Update patch set 4
Patch Set 4:

Damn, the new CQ is just too fast. I was going to comment on this yesterday and wanted to test it out first, but then got stuck in 2 hours of infra issues. :/

Anyway, do we really need to dump a mountain of preprocessor magic in here for this? Half of that is even checking for __cplusplus which I think should never apply to vboot. Can't we just use __attribute((fallthrough)) directly in the code? Sounds like that's supported by both GCC and clang. (Or ideally we could hold back on clang updates that copy a long-standing GCC warning until they actually manage to copy the full behavior, which it sounds like they're only planning for 3.2, but I guess that's not an option here?)

Patch-set: 4
CC: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>