cgpt: improve help text
Standardize on the term "blocks" rather than "sectors" as this matches
our disk layout configuration files.

For commands that operate on blocks, clarify the unit assumptions (in
that 1 block is 512 bytes).

For cgpt show, split the specific field options out from the -i option.
The current text makes it seem like they're always required when using
the -i option, and it's not super clear that they only matter when the
-i flag is used.

BUG=None
TEST=CQ passes
BRANCH=None

Change-Id: Idfd939cfe9501da78e323bc1ce8c1fbf772fbaa9