Update patch set 4
Patch Set 4:

> I will remove the c++ specific code. Note: Current versions of both gcc and clang in the chroot do not support the attribute. So, we can't  use the attribute directly :(.

Well, I assume clang is going to support it soon (after your uprev)? The old CrOS GCC doesn't support it, but it currently doesn't build vboot anyway (see Mike's comment in crbug.com/991812)... I'm not sure if we even want to try to support that anymore because it's so old. Maybe we can find a better way to work around that problem (will continue that on the bug).

> Regarding copying full behavior, I am not sure whether suppressing the warning based on fallthrough string in a comment is currently possible.

Well, GCC does this, and even documents its exact behavior in nice detail (https://gcc.gnu.org/onlinedocs/gcc-8.3.0/gcc/Warning-Options.html#Warning-Options). I found a hint here that clang is intending to do this in a later version: https://stackoverflow.com/questions/8809154/how-to-make-gcc-clang-warn-about-missing-breaks-in-switch-statements . (In my opinion it's a bad decision from the clang team to add this warning with the same name if they're not going to match GCC's behavior, even for one version... but I guess that's not really relevant to the question what we're going to do about it now.)

Patch-set: 4
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>