Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I hope these groot-specific functions and files are only temporary?

Hi Joel,

Yes, EVERYTHING is temporary right now as we are still prototyping and everything has been marked "WIP".  These are just intermediate CLs so that Ting and I can  share code and stay in sync.  There's a lot of cleanup involved before these CLs are supposed to be seen by anyone outside of the two of us :p.

Patch-set: 3