Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> What's the reason for implementing a stack here?  Do we now have multiple paths to get to the same menu?  If so, would remembering the "last menu" be sufficient?

Yeah....so the problem is that our history has gotten longer.  For example, from each screen, we can select an "Advanced Options" menu item, and in the Advanced Options Menu, we have selections for "BIOS log" or "Debug Options".  So, going back to the Advanced Options Menu is no problem, but we need to let the user return to the original screen where they selected Advanced Options from, right?

Patch-set: 3