Update patch set 2
Patch Set 2:

Using a model has the disadvantage that we have to keep two files in sync: as soon as we change the behavior here, we'd have to remember to update the model file. The approach I now use is better than the function-level comment because it is supposed to clear the tainted flag only on successful parsing. Should, because that part of Coverity isn't very well documented, at least in public

Patch-set: 2