firmware: Do not set recovery reason directly in LoadKernel()
LoadKernel() currently contains code that sets the recovery reason
directly (via direct nvdata access, bypassing the usual
VbSetRecoveryReason() helper) whenever it has a problem loading a
kernel. This seems to be an ancient vestige from the time when
LoadKernel() (and not VbSelectAndLoadKernel()) was still the external
API. In our current use, VbTryLoadKernel() will always immediately
override any recovery reason set this way.

This patch removes this pointless code to avoid confusion. Ultimately,
with the full refactoring of kernel verification we will probably want
to get to a system where the more precise recovery reasons from loading
the kernel (e.g. "no kernel" vs "invalid kernel") can be carried through
to the outer layer. But that's not how the code works right now and it
would be a larger endeavor to move in that direction, so let's make sure
our code is at least consistent (and doesn't do things that always
immediately get undone again) for now and leave that aspect for later
larger-scale refactoring.

BRANCH=None
BUG=None
TEST=make runtests

Change-Id: Idd8bd6e16d5ef1472aa3b2b66468248726d5c889
Signed-off-by: Julius Werner <jwerner@chromium.org>