ec_sync: Ditch VbExEcGetExpectedImage callback
This callback is redundant; the underlying platform is expected to
hand over a buffer and size to this callback, only to get passed those
same arguments to VbExEcUpdateImage.  The platform has to know where
the FW is located, so let it handle that itself in VbExEcUpdateImage.

VbExEcUpdateImage was therefore updated to remove the buffer and size
arguments.

BUG=none
BRANCH=none
TEST=make clean && make runtests

Change-Id: I3dc3f8aa28c16a430aa348e3a13a417731e44ae2
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>