Always initialize CGPT entries buffers.
The current code only initializes the buffer in AllocAndReadGptData
when the secondary header is valid, but then goes a head and checks
the secondary entries unconditionally in GptSanityCheck.

This makes the fuzzer happy again, but not sure this is the correct
fix. Probably we should rather teach GptSanityCheck with a situation
where the entries buffers are invalid?

BUG=chromium:1017797
TEST=See fuzzer testcase in bug.
BRANCH=none

Change-Id: Ib21492f6d563330fc8c9897288276657164b6c78