Update patch set 2
Patch Set 2:

(1 comment)

> Currently the goal is to converge on 2lib or vboot2, which means we should actually be using vboot_fw2x.a.

I think it would be best to just have a single vboot_fw.a with everything in it to be honest. See the other discussion about this happening in CB:36341. The only remaining problem is that we have to divorce the Tlcl parts from vboot a bit, otherwise we get multiple definition errors from the stuff that was copied from it into coreboot.

Patch-set: 2