battery cutoff: Move cutoff handling from EC sync to kernel load
Currently, battery cutoff is handled at the end of a successful EC
software sync.  Now that auxiliary firmware sync is separate from the
EC, this patch moves it to the end of kernel select and load, which
means that auxiliary firmware will get updated as well before battery
cutoff.

BUG=none
BRANCH=none
TEST=make runtests

Change-Id: I96bea22667ebf45b446a26d84de96e52f3d289a5
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>