battery cutoff: Move cutoff handling from EC sync to kernel load
Currently, battery cutoff is handled at the end of a successful EC
software sync.  Now that auxiliary firmware sync is separate from the
EC, this patch moves it back to after both EC and auxfw updates are
successful, to ensure all firmware is up-to-date before entering ship
mode.

BUG=none
BRANCH=none
TEST=make runtests

Change-Id: I96bea22667ebf45b446a26d84de96e52f3d289a5
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>