Update patch set 3
Patch Set 3:

> Normally, update_chroot is all that you need. Can you try with ./update_chroot --nousepkg ?
> Warning: Building llvm/clang locally will take over an hour or more.

Yes, running with --nousepkg makes it start building the new version. (I didn't let it finish because I didn't want to wait an hour.)

Can you figure out how to fix it so that it also works correctly with prebuilds? We can't ask the whole team to build this from source.

Patch-set: 3