Update patch set 2
Patch Set 2:

This doesn't make sense.  I think this needs to be debugged on the depthcharge side.  The only thing that auxfw_finalize does in depthcharge is to disable the I2C "tunnels" that are used to update the TCPC firmware, which should be done regardless if the TCPCs were updated or not.

Patch-set: 2
Reviewer: Gerrit User 1327490 <1327490@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>