Update patch set 1
Patch Set 1:

Subrata, right now this flag is guarded in the Makefile by FIRMWARE_ARCH being i386 (i.e., the host tools).  If we want to enable this on the DUTs, then depthcharge can take on the ability to define this flag or not.

Patch-set: 1
Reviewer: Gerrit User 1327490 <1327490@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>