Update patch set 1
Patch Set 1:

Not like this, please. If you want to re-enable it on x86, please fix the bug in the Makefile instead (the bug is that it checks for 'i386' when it should check for 'x86' or 'x86_64'). If you want to enable it for all platforms, please prove that it actually speeds things up on Arm first.

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>