futility: updater: move "EC partial recovery" to be a quirk.
The "EC update" (either calling flashrom or using EC RO software sync)
logic has been bloated and is really not a typical updater feature (that
only makes sense for dogfooders). And we have seen enough special cases
that some boards may not want to use it (or causing problems).

Move that to a quirk so we can turn on or off in a more flexible way.

BRANCH=none
BUG=chromium:1024401
TEST=make clean && make runtests

Change-Id: I2d4fe0d9ee0d98ad41b8cbdcaff848846d120d07