vboot: Don't warn user if enter key is pressed if rec switch is physical
When the recovery switch is physical then the prompt to enter dev mode
will say to press the recovery switch not the enter key. In this case we
do not want to warn the user that an internal keyboard is needed to
confirm we will just silently ignore the enter key.

BUG=b:144034020
TEST=On Drallion enter key does not work to switch to dev mode and does
     not beep or print a warning when the enter key is pressed
BRANCH=none

Change-Id: I250ea2622c9c38bfc0d7463eb95ca65a12b41153