Update patch set 1
Patch Set 1:

You will need to also remove the code in autotest that sets this GBB flag.

But I also take issue with removing it at all.  Sure we'd all like to "avoid similar workarounds in the future" but the reality is we do not live in a perfect world and this has been used a few crucial times to ship a platform so don't be surprised if it has to be resurrected.

Patch-set: 1
CC: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>