Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1: Commit-Queue+2
> 
> > > But I also take issue with removing it at all.  Sure we'd all like to "avoid similar workarounds in the future" but the reality is we do not live in a perfect world and this has been used a few crucial times to ship a platform so don't be surprised if it has to be resurrected.
> > 
> > I think our line of thought was that if a product needs different behavior, it would be cleaner to modify the test than modify the firmware (e.g. create a new FAFT config that is set for that board and makes it skip or change a particular test). Do you think that would work?
> 
> I'm also a little skeptical of claiming that a particular machine "passes" a FAFT test if it has to enable this GBB flag in order to do so.  Seems like it would make more sense to "waive" the test (CTS-style) or skip/modify the test based on FAFT config as Julius suggested.

I agree with the sentiment to be sure, but the reality is the cases that I used this for would not have been possible to waive or hack into tests themselves.

I'm ok removing this for now because it was being abused from it's intended purpose.  But I'll also be ok approving a similar hack in the future if we need something to ship a device.

Patch-set: 1
Reviewer: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2