vboot/tlcl: remove TlclSetGlobalLock function from tpm2_lite
In TPM2, TlclWriteLock is used instead of TlclSetGlobalLock
to lock secdata spaces.

Although the function is defined in the common tlcl.h header,
it does not have to be implemented in the tpm2_lite/tlcl.c
code.  This is what coreboot does.

Add a compile-time conditional in utilility/tpmc.c to leave
out the TlclSetGlobalLock function for TPM2.

BUG=b:124141368, chromium:972956
TEST=make clean && make runtests
BRANCH=none

Change-Id: I99bb04bbc2f4eef28cbe51ecf506078583238074
Signed-off-by: Joel Kitching <kitching@google.com>