futility: fix phaser360 dual root key issue
Because of lacking CL:1501614 in octopus factory branch, dopefish
root key is written to phaser360 model at a time segment. That
will lead to firmware updater not be able to verify RW vblock and
AU will fail. This CL will fix that by using root key info and
model name to make firmware updater get a proper sig_id so that
in-field machine can be updated by AU.

Change-Id: I163c16189c28a996ed08bf2a7b162e6ee3b13be6
Signed-off-by: magf@bitland.corp-partner.google.com