futility: updater: fix phaser360 dual root key issue
Because of lacking CL:1501614 in octopus factory branch, dopefish
root key is written to phaser360 model at a time segment. That
will lead to firmware updater not be able to verify RW vblock and
AU will fail. This CL will fix that by using root key info and
model name to make firmware updater get a proper sig_id so that
in-field machines can be updated by AU.

BUG=b:146876241, b:133901651, b:146482979
BRANCH=none
TEST=using a DUT of phaser360 (without whitelabel_tag = dopefish)
     which is flashed dopefish rootkey and hwid, using command
     'chromeos-firmwareupdate -m autoupdate --wp=1' to flash
     firmware, RW firmware can be updated and DUT can boot normally.

Change-Id: I163c16189c28a996ed08bf2a7b162e6ee3b13be6
Signed-off-by: Paul Ma <magf@bitland.corp-partner.google.com>