Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is intentional so that the Pre-CQ can catch when a patch breaks the fuzzers. Any reason you need to change this?

Thanks. Fuzzer build is breaking the instrumentation of llvm and we are not going to have this patch checked in. For now keeping this patch so that we can launch tryjobs to work around this issue.

Patch-set: 1
Reviewer: Gerrit User 1148807 <1148807@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>