Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > It might be a good idea to update the boards on coreboot.org that disable EC software sync by default to also select this GBB flag so their behavior does not change.
> 
> The Hatch family is the only case where EC sync is done in coreboot right now, and AUXFW sync isn't supported in coreboot right now, either.  AFAICT, this would only affect depthcharge.
> 
> We already added a CL to add CONFIG_AUXFW_SOFTWARE_SYNC to depthcharge for all devices that were using EC sync before the two were split up.

This was more about the GBB flags because the selected flags in the mainboard Kconfig can determine the default state of the firmware in the factory.  It looks like only Sarien/Drallion/Trogdor set the GBB flag to disable software sync by default but that was implicitly disabling PD sync as well.

Patch-set: 3