vboot/vboot_display: retrieve TPM versions from vboot1 structures
The sd->kernel_version_secdatak variable referenced here is never
actually initialized by current code, so it always shows as 0x00000000.
This patch reverts part of the changes regarding TPM versions in
CL:1680190.

This CL is a temporary fix for kukui. Please do not merge this CL to
ToT, which will be fixed by initializing the variable in question
instead (see CL:1651221).

BRANCH=kukui
BUG=none
TEST=emerge-kukui vboot_reference

Change-Id: I86316a175be7a29eb9c0e4e67dfd59c0ddff9fed