Update patch set 1
Patch Set 1:

Also, what should we do for the cases where an assertion is more suitable, but we still want it to be skipped in the case of recovery mode?

For example -- for vb2_nv_get() and vb2_nv_set(), we currently need to use either:

  VB2_ASSERT((ctx->flags & VB2_CONTEXT_RECOVERY_MODE) ||
             vb2_get_sd(ctx)->status & VB2_SD_STATUS_NV_INIT)

  if (!(vb2_get_sd(ctx)->status & VB2_SD_STATUS_NV_INIT))
      VB2_REC_OR_DIE(ctx, "nvstorage not initialized\n");

Would we want to also create a VB2_REC_OR_ASSERT for these cases?
Or do we want to change the structure of vb2_nv_get/set functions to be more similar to vb2_secdata_get and friends?

Or is it possible to create a helper macro for the recovery mode checking?  E.g.

  VB2_REC_OR(VB2_DIE(...))
  VB2_REC_OR(VB2_ASSERT(...))

Or do we just want a helper function vb2_is_rec(ctx)?

  vb2_is_rec(ctx) || VB2_DIE(...)
  vb2_is_rec(ctx) || VB2_ASSERT(...)

Patch-set: 1
Reviewer: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>