Update patch set 1
Patch Set 1:

(3 comments)

> Also, what should we do for the cases where an assertion is more suitable, but we still want it to be skipped in the case of recovery mode?
> 
> Would we want to also create a VB2_REC_OR_ASSERT for these cases?

For now it seems like in practice we preferred to use VB2_DIE() and write a custom error message, so I was just catering to that case for now. I think in practice the difference might be that in cases where you want to ignore an abort in recovery mode, you usually still have to do something (e.g. return 0) in that path, so you can't easily do that with a macro where both the check and the abort are rolled into a single macro.

> Or do we want to change the structure of vb2_nv_get/set functions to be more similar to vb2_secdata_get and friends?

Yes, I think we should.

> Or is it possible to create a helper macro for the recovery mode checking?  E.g.
> 
>   VB2_REC_OR(VB2_DIE(...))
>   VB2_REC_OR(VB2_ASSERT(...))
> 
> Or do we just want a helper function vb2_is_rec(ctx)?
> 
>   vb2_is_rec(ctx) || VB2_DIE(...)
>   vb2_is_rec(ctx) || VB2_ASSERT(...)

Both of these options look way more cluttery than my macro to me. Let's not overcomplicate things, we don't have that many abort situations that we need a fully customizable framework for it.

Patch-set: 1