Update patch set 2
Patch Set 2:

> Where is -o pipefail being set?

At the top of this script? (I don't know if there's anything here really needing that, but I don't want to spend the time finding out, and generally it's something you probably want to have if you're using pipes and relying on return values a lot.)

> Vapier said he was able to reproduce the problem (?), do we have a way of testing to make sure this fixes it?

I think submitting this and checking if the flakes disappear is the easiest test.

Patch-set: 2