Update patch set 14
Patch Set 14:

> Actually, I'm not sure why one would need to run set_gbb_flags.sh in an SDK build?  The script seems hard-coded to use `flashrom -p host`, so it should only work on a device, no?

It has a -f flag to work on a file and I think people do that on the host frequently. (I wouldn't mind someone implementing a -p flag so it can be used over servo as well...)

Maybe this is fine, if nobody else complains we can just leave it at this and people can get used to the new path. Definitely an improvement that invocation is no longer different between host and DUT. Just wasn't really aware this patch was doing that.

Patch-set: 14