Update patch set 14
Patch Set 14:

> Actually, I think there's already an official replacement for the "-f" use case -- `futility gbb_utility bios_file`.  I don't think it can list GBB flags like set_gbb_flags.sh does when invoked without any arguments, and I'm not sure if it can operate on the current host's flash memory directly?  We might consider ensuring it covers all use cases and recommend that solution rather than continuing to support two.

I mean, set_gbb_flags.sh is just a wrapper around gbb_utility. It does things like print documentation and saves you the effort of which flags to pass to gbb_utility around this most common of use cases. I agree using it with -f isn't a very big difference to running the utility directly, but I think people still like to use it (also because it works similar to the way they're used to running it on the DUT).

Patch-set: 14