Update patch set 1
Patch Set 1:

I know you said you prefer it if we skipped this step... but I was creating a helper function to perform the conversion, and it seemed pretty trivial to convert it to an API function.

It also unblocks us from removing vboot1 data structures from depthcharge (with the sole exception of crossystem_*.c).

Patch-set: 1