futility: updater: Ignore image parsing error in --force
The firmware updater has been improved to support flashing outside DUT
(--ccd, --servo) that also implies more people will use it for devices
with corrupted (or empty) firmware. It's pretty confusing for developers
to see "Cannot load system active firmware" while the flashrom can
actually read and write to the SPI firmware.

The solution here is to allow updating on such devices when --force is
specified (which is automatically applied for --servo and --ccd).

BUG=b:148405957
TEST=make runtests
BRANCH=None

Change-Id: I19e63e3464616bc508639cbfad0d1cf8e99507b0
Signed-off-by: Hung-Te Lin <hungte@chromium.org>