vboot: Add mockable attribute to some over-ridden functions
- Test only, do not review-

A newer clang version has an optimization that is breaking
vboot_reference unit tests. The root cause appears to be:
Many functions inside unit tests are supposed to override
the functions defined in the main library but the main
library functions are not defined as weak.

Fix by marking some of the overridden function by
test_mockable which is already used for some of the functions.

This is similar to the prior bug chromium:723906 where unit tests
were broken after binutils upgrade.

BUG=chromium:1055151
TEST=Pending

Change-Id: Ib9f16e1af59632b6a82411656ebba75e26b5d211