Update patch set 1
Patch Set 1:

> Hmmm, yeah, honestly... is this coming from any specific need? I don't really see the big deal about providing an empty VbExDisplayScreen() (or however the new API is gonna be called) in coreboot that would make it worth overhauling the whole EC sync stuff yet again (I guess I'm just getting tired of it, we've done it so often recently). If you don't like that another simple solution would be to provide a compile-time EC_SYNC_CAN_DISPLAY=1 switch that depthcharge passes but coreboot doesn't. This just feels like a big change with little incentive, and we have so many other pressing things that would need attention in vboot after all.

The original impetus for this is that we are creating a new API for Groot to use, and leaving VbExDisplayScreen to be deprecated.  The way Yu-Ping currently has this set up in depthcharge is to conditionally compile one of the UI files, depending on which UI flag is selected.  So this means that (1) vboot EC sync code needs to conditionally call the correct UI depending on the flag, (2) coreboot needs to implement all possible UI functions, even though they are empty.

Since we need to touch coreboot code anyways, I figured it wouldn't be much more work to get rid of EC/auxfw sync's dependency on UI altogether.  That also means we don't need so many conditionals around the UI flag sprinkled throughout vboot_reference code.

I guess we can do the EC_SYNC_CAN_DISPLAY=1 thing, but I think overall this is doing some more work up front to get everything in a better state (removing more dependencies on old UI and vboot1 stuff, preparing to deprecate old UIs, modularizing UI code).

If we move to the one-function solution, then it'll be much less code changes in 2ec_sync.c; would that be better?

Patch-set: 1
Reviewer: Gerrit User 1126709 <1126709@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>