Update patch set 1
Patch Set 1:

> I guess we can do the EC_SYNC_CAN_DISPLAY=1 thing, but I think overall this is doing some more work up front to get everything in a better state (removing more dependencies on old UI and vboot1 stuff, preparing to deprecate old UIs, modularizing UI code).

I mean, at the end of the day we still have a single thing that we want to happen in one shot (without the calling firmware needing to understand the individual stages and decisions) which needs to do something platform dependent (display an image) in the middle. It feels to me like a callback is now and in the future the simplest solution for that. That doesn't prevent us from modularizing UI, it just means that rather than calling VbExDisplayScreen() or whatever it would call a single-purpose vb2ex_ec_sync_slow_warning() callback that leaves all the UI details of what exactly that warning is to the calling firmware.

> If we move to the one-function solution, then it'll be much less code changes in 2ec_sync.c; would that be better?

Yes, I think in terms of complexity that would be better than what this patch has. But I still feel that just calling a callback is even simpler than running the function twice.

Patch-set: 1