Makefile: Explicitly signal success at the end of 'make runtests'
If you run 'make runtests' with -j (as you should because it takes
forever), it can be hard to see if it passed successfully. A 'make: ***
[Makefile:XXX: ...] Error 255' line might hide somewhere in the middle
with a lot of successful output from other tests running in parallel
hiding it.

It's hard (I think?) to have make output something at the end (after
joining all processes) when it has encountered an error, but at least it
is easy to output something when the 'runtests' target finishes
successfully. This patch adds such a highly visible success message, so
that its absence will make it clear that there was an error further up.

BRANCH=None
BUG=None
TEST='make runtests', both passing and failing

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: If979a7635f5c05c44a92daca12d31e344563794c