crossystem: use GPIO chardev API instead of sysfs
We used to read GPIO state from sysfs which caused
trouble due to pin offsets changing in new kernel versions.

Kernel v4.8 introduced a GPIO character device API where offsets
are supposed to be stable with respect to gpiochip base.

Use raw ioctls instead of libgpiod as it is best to avoid adding
unnecessary dependencies to vboot_reference.

Fall back to sysfs if the new method fails or is not supported.

BUG=b:143174998
BRANCH=none
TEST=crossystem wpsw_cur (on Kefka)

Signed-off-by: Michal Stanek <mst@semihalf.corp-partner.google.com>
Signed-off-by: Lukasz Majczak <lma@semihalf.com>
Change-Id: I67be7875173ae60adb8a5d5f059c2d91b76d7246